feat(town): add tickTownFollowers module and wire into runtime to manage town followers
