+++ ROguelike 3_10_2025 filut jarjestyksessa luultavasti vikaa/core/town/runtime.js
@@ -14,6 +14,7 @@
 import { getMod } from "../../utils/access.js";
 import { syncFollowersFromTown } from "../followers_runtime.js";
 import { tickSeppo } from "./seppo_runtime.js";
+import { tickTownFollowers } from "./follower_tick.js";
 
 export function generate(ctx) {
   // Ensure townBiome is not carrying over from previous towns; allow derive/persist per town
@@ -1155,80 +1156,7 @@ export function tick(ctx) {
   } catch (_) {}
 
   // Simple follower NPC behavior: stay near the player in town, unless set to wait.
-  try {
-    const p = ctx.player;
-    if (p && Array.isArray(ctx.npcs)) {
-      const followers = p && Array.isArray(p.followers) ? p.followers : null;
-      for (const n of ctx.npcs) {
-        if (!n || !n._isFollower) continue;
-
-        // Resolve follower mode from the record (or NPC override) so town followers
-        // can obey simple follow / wait commands.
-        let mode = "follow";
-        try {
-          if (n._followerMode === "wait" || n._followerMode === "follow") {
-            mode = n._followerMode;
-          } else if (followers && n._followerId != null) {
-            const rec = followers.find(f => f && f.id === n._followerId) || null;
-            if (rec && (rec.mode === "wait" || rec.mode === "follow")) {
-              mode = rec.mode;
-            }
-          }
-        } catch (_) {}
-        if (mode === "wait") continue;
-
-        const dx = p.x - n.x;
-        const dy = p.y - n.y;
-        const dist = Math.abs(dx) + Math.abs(dy);
-        const followRange = 2;
-        if (dist <= followRange) continue;
-
-        const sx = dx === 0 ? 0 : (dx > 0 ? 1 : -1);
-        const sy = dy === 0 ? 0 : (dy > 0 ? 1 : -1);
-        const primary = Math.abs(dx) > Math.abs(dy)
-          ? [{ x: sx, y: 0 }, { x: 0, y: sy }]
-          : [{ x: 0, y: sy }, { x: sx, y: 0 }];
-
-        let moved = false;
-        for (const d of primary) {
-          const nx = n.x + d.x;
-          const ny = n.y + d.y;
-          if (isFreeTownFloor(ctx, nx, ny)) {
-            if (ctx.occupancy && typeof ctx.occupancy.clearNPC === "function") {
-              ctx.occupancy.clearNPC(n.x, n.y);
-            }
-            n.x = nx;
-            n.y = ny;
-            if (ctx.occupancy && typeof ctx.occupancy.setNPC === "function") {
-              ctx.occupancy.setNPC(n.x, n.y);
-            }
-            moved = true;
-            break;
-          }
-        }
-        if (!moved) {
-          const ALT_DIRS = [
-            { x: 1, y: 1 }, { x: 1, y: -1 }, { x: -1, y: 1 }, { x: -1, y: -1 }
-          ];
-          for (const d of ALT_DIRS) {
-            const nx = n.x + d.x;
-            const ny = n.y + d.y;
-            if (isFreeTownFloor(ctx, nx, ny)) {
-              if (ctx.occupancy && typeof ctx.occupancy.clearNPC === "function") {
-                ctx.occupancy.clearNPC(n.x, n.y);
-              }
-              n.x = nx;
-              n.y = ny;
-              if (ctx.occupancy && typeof ctx.occupancy.setNPC === "function") {
-                ctx.occupancy.setNPC(n.x, n.y);
-              }
-              break;
-            }
-          }
-        }
-      }
-    }
-  } catch (_) {}
+  tickTownFollowers(ctx);
 
   // Rebuild occupancy every other turn to avoid ghost-blocking after NPC bursts
   try {
